Unit tests

+++ services/userService.ts
@@ -84,8 +84,9 @@ export const phoneChange = async ({ userId, phone }: IChangePhone) => {
                 phone,
                 `Hello ${phoneUpdate.firstName}, Verify your phone number by clicking ${link}`
             );
-            await phoneChangeNotification.forced_send();
+            const phoneID = await phoneChangeNotification.forced_send();
 
+            logInfo(phoneID.res);
             logInfo(`Phone change link : ${link}`);
             logInfo(`Phone notification has been sent`);
 
@@ -115,7 +116,10 @@ export const deleteUserData = async ({ userId }: any) => {
     const AuthData = await getUserAuthData(userId);
     let transactionArray = [];
 
-    if (user != null && AuthData != null && user.phone != null) {
+    if (user != null && AuthData != null) {
+        // If the user registers for the first time,
+        // they dont give their phone numbers and then decide to delete the account,
+        // the check on phone number would always fail. The transaction array later has been modified too.
         const deletePupilData = prisma.pupil.deleteMany({
             where: {
                 userId: userId,
@@ -151,19 +155,23 @@ export const deleteUserData = async ({ userId }: any) => {
                 id: userId,
             },
         });
-
-        // we are using above user.phone != null because we don't need phone while registering the user, and if user wants to delete his/her account before completing his data(phone) we need that null otherwise below query expecting a string value over a null value from DB
-        const deleteTextData = prisma.textNotifications.deleteMany({
-            where: {
-                recipientPhone: user.phone,
-            },
-        });
+        // we are using above user.phone != null because we don't need phone while registering the user,
+        // and if user wants to delete his/her account before completing his data(phone)
+        // we need that null otherwise below query expecting a string value over a null value from DB
+        let deleteTextData = undefined;
+        if (user.phone !== null) {
+            deleteTextData = prisma.textNotifications.deleteMany({
+                where: {
+                    recipientPhone: user.phone,
+                },
+            });
+        }
 
         transactionArray = [
             deletePupilData,
             deleteVolunteerData,
             deleteEmailData,
-            deleteTextData,
+            ...(deleteTextData !== undefined ? [deleteTextData] : []),
             deleteRefreshTokens,
             deleteAuthData,
             deleteUser,
@@ -196,7 +204,7 @@ export const userUpdate = async (userData: any, userId: string) => {
 
         if (dataKeys.includes('email')) {
             if (findUser.email != userData['email']) {
-                logInfo('Sending verification email yo updated email');
+                logInfo('Sending verification email to updated email');
 
                 await prisma.user.update({
                     where: {
